API

Scheduler Control

JobSchedulers.scheduler_status — Function

scheduler_status() :: Symbol

Print the settings and status of job scheduler. Return :not_running or :running.

JobSchedulers.scheduler_start — Function

scheduler_start()

Start the job scheduler.

JobSchedulers.scheduler_stop — Function

scheduler_stop()

Stop the job scheduler.

Scheduler Settings

JobSchedulers.set_scheduler — Function

set_scheduler(;
    max_cpu::Union{Int,Float64} = JobSchedulers.SCHEDULER_MAX_CPU,
    max_mem::Union{Int,Float64} = JobSchedulers.SCHEDULER_MAX_MEM,
    max_job::Int = JobSchedulers.JOB_QUEUE_MAX_LENGTH,
    update_second = JobSchedulers.SCHEDULER_UPDATE_SECOND
)

See details: set_scheduler_max_cpu, set_scheduler_max_mem, set_scheduler_max_job, set_scheduler_update_second

JobSchedulers.set_scheduler_max_cpu — Function

set_scheduler_max_cpu(ncpu::Int = default_ncpu())
set_scheduler_max_cpu(percent::Float64)

Set the maximum CPU (thread) the scheduler can use. If starting Julia with multi-threads, the maximum CPU is nthreads() - 1.

Example

set_scheduler_max_cpu()     # use all available CPUs
set_scheduler_max_cpu(4)    # use 4 CPUs
set_scheduler_max_cpu(0.5)  # use 50% of CPUs

JobSchedulers.set_scheduler_max_mem — Function

set_scheduler_max_mem(mem::Int = default_mem())
set_scheduler_max_mem(percent::Float64)

Set the maximum RAM the scheduler can use.

Example

set_scheduler_max_mem()             # use 80% of total memory

set_scheduler_max_mem(4GB)          # use 4GB memory
set_scheduler_max_mem(4096MB)
set_scheduler_max_mem(4194304KB)
set_scheduler_max_mem(4294967296B)

set_scheduler_max_mem(0.5)          # use 50% of total memory

JobSchedulers.set_scheduler_update_second — Function

set_scheduler_update_second(s::Float64 = 0.6)

Set the update interval of scheduler.

JobSchedulers.set_scheduler_max_job — Function

set_scheduler_max_job(n_finished_jobs::Int = 10000)

Set the number of finished jobs.

Job

JobSchedulers.Job — Type

Job(command::Base.AbstractCmd; stdout=nothing, stderr=nothing, append::Bool=false, kwargs...)
Job(f::Function; kwargs...)
Job(task::Task; kwargs...)

Arguments

command::Base.AbstractCmd: the command to run.

f::Function: the function to run without any arguments, like f().

task::Task: the task to run. Eg: @task(1+1).

Common Keyword Arguments (kwargs...)

name::String = "": job name.

user::String = "": user that job belongs to.

ncpu::Int64 = 1: number of CPU this job is about to use.

mem::Int64 = 0: number of memory this job is about to use (supports TB, GB, MB, KB, B=1).

schedule_time::Union{DateTime,Period} = DateTime(0): The expected time to run.

dependency: defer job until specified jobs reach specified state (QUEUING, RUNNING, DONE, FAILED, CANCELLED, PAST). PAST is the super set of DONE, FAILED, CANCELLED, which means the job will not run in the future. Eg: DONE => job, [DONE => job1; PAST => job2].

To be compatible with old versions, you can also use job id (Int): [DONE => job.id].

wall_time::Period = Week(1): wall clock time limit.

priority::Int = 20: lower means higher priority.

Experimental Keyword Arguments - Output Redirection:

stdout=nothing: redirect stdout to the file.

stderr=nothing: redirect stderr to the file.

append::Bool=false: append the stdout or stderr or not.

Note

Redirecting in Julia are not thread safe, so unexpected redirection might be happen if you are running programs in different Tasks simultaneously (multi-threading).

JobSchedulers.submit! — Function

submit!(job::Job)

Submit the job. If job.create_time == 0000-01-01T00:00:00 (default), it will change to the time of submission.

JobSchedulers.cancel! — Function

cancel!(job::Job)

Cancel job, stop queuing or running.

JobSchedulers.result — Function

result(job::Job)

Return the result of job. If the job is not done, a warning message will also show.

Queue

JobSchedulers.queue — Function

queue(; all::Bool = false)    -> Vector{Job}
queue(state::Symbol )         -> Vector{Job}
queue(needle)                 -> Vector{Job}
queue(state::Symbol , needle) -> Vector{Job}
queue(needle, state::Symbol ) -> Vector{Job}
queue(id::Int)                -> Job

all::Bool: if true, get all jobs. if false, get only running and queuing jobs.

state::Symbol: get jobs with a specific state, including :all, QUEUING, RUNNING, DONE, FAILED, CANCELLED, PAST.

PAST is the superset of DONE, FAILED, CANCELLED.

needle::Union{AbstractString,AbstractPattern,AbstractChar}: get jobs if they contain needle in their name or user.

id::Int: get the job with the specific id.

JobSchedulers.all_queue — Function

all_queue()
all_queue(id::Int64)
all_queue(state::Symbol)
all_queue(needle::Union{AbstractString,AbstractPattern,AbstractChar})

state::Symbol: get jobs with a specific state, including :all, QUEUING, RUNNING, DONE, FAILED, CANCELLED, PAST.

PAST is the superset of DONE, FAILED, CANCELLED.

needle::Union{AbstractString,AbstractPattern,AbstractChar}: get jobs if they contain needle in their name or user.

id::Int: get the job with the specific id.

JobSchedulers.job_query — Function

job_query_by_id(id::Int64)

Search job by job.id in the queue.

Return job::Job if found, nothing if not found.

Wait For All Jobs

JobSchedulers.wait_queue — Function

wait_queue(;show_progress = false)

Wait for all jobs in queue() become finished.

show_progress = true, job progress will show.

See also: queue_progress.

Optimize CPU Usage

JobSchedulers.solve_optimized_ncpu — Function

solve_optimized_ncpu(default::Int; 
    ncpu_range::UnitRange{Int64} = 1:total_cpu, 
    njob::Int = 1, 
    total_cpu::Int = JobSchedulers.SCHEDULER_MAX_CPU, 
    side_jobs_cpu::Int = 0)

Find the optimized number of CPU for a job.

default: default ncpu of the job.

ncpu_range: the possible ncpu range of the job.

njob: number of the same job.

total_cpu: the total CPU that can be used by JobSchedulers.

side_jobs_cpu: some small jobs that might be run when the job is running, so the job won't use up all of the resources and stop small tasks.

Backup

JobSchedulers.set_scheduler_backup — Function

set_scheduler_backup(
    filepath::AbstractString = "";
    migrate::Bool = false,
    delete_old::Bool = false,
    recover_settings::Bool = true,
    recover_queue::Bool = true
)

Set the backup file of job scheduler.

If filepath was set to "", stop backup at exit.

If filepath was set to an existing file, recover_settings or recover_queue from filepath immediately.

If filepath was set to a new file, the backup file will be created at exit.

If migrate=true and the old JobSchedulers.SCHEDULER_BACKUP_FILE exists, the old backup file will be recovered before recovering from filepath.

JobSchedulers.backup — Function

backup()

Manually backup job scheduler settings and queues. The function is automatically triggered at exit.

Progress Meter

Note

To display a progress meter, please use wait_queue(show_progress = true).

JobSchedulers.JobGroup — Type

mutable struct JobGroup
    total::Int
    queuing::Int
    running::Int
    done::Int
    failed::Int
    cancelled::Int
    eta::Millisecond
    group_name::String
    job_name::String
    failed_job_names::Vector{String}
    elapsed_times::Vector{Millisecond}
    function JobGroup(group_name)
        new(0, 0, 0, 0, 0, 0, Millisecond(0), group_name, "", String[], Millisecond[])
    end
end

JobGroup is computed when displaying a progress meter.

JobSchedulers.fingerprint — Function

fingerprint(g::JobGroup)
ALL_JOB_GROUP_FINGERPRINT = fingerprint(ALL_JOB_GROUP)

If ALL_JOB_GROUP_FINGERPRINT is different from new fingerprint, update view.

JobSchedulers.get_group — Function

get_group(job::Job, group_seperator = r": *")

Return nested_group_names::Vector{String}.

Eg: If job.name is "A: B: 1232", return ["A", "A: B", "A: B: 1232"]

JobSchedulers.progress_bar — Function

progress_bar(percent::Float64, width::Int = 20)

Return ::String for progress bar whose char length is width.

percent: range from 0.0 - 1.0, or to be truncated.

width: should be > 3. If <= 10, percentage will not show. If > 10, percentage will show.

JobSchedulers.queue_progress — Function

queue_progress(;remove_tmp_files::Bool = true, kwargs...)
queue_progress(stdout_tmp::IO, stderr_tmp::IO;
group_seperator = r": *", wait_second_for_new_jobs::Int = 1, loop::Bool = true)

group_seperator: delim to split (job::Job).name to group and specific job names.

wait_second_for_new_jobs: if auto_exit, and all jobs are PAST, not quiting queue_progress immediately but wait for a period. If new jobs are submitted, not quiting queue_progress.

loop: if false, only show the current progress and exit.

JobSchedulers.view_update — Function

view_update(h, w; row = 1, groups_shown::Vector{JobGroup} = JobGroup[], is_in_terminal::Bool = true, is_interactive = true, group_seperator_at_begining = r"^: *")

Update the whole screen view.